Add Julia version-independent resolution capability
In the BinaryBuilder world, we use `Pkg.add()` to install dependent JLLs
in the build prefix.  By relying on the Pkg resolver, we implicitly
limit ourselves to the package versions that are compatible with the
currently-running Julia.  This prevents us from, for example, using
Julia 1.6 in BB to generate packages that link against the binaries used
by Julia 1.5, or from constructing "impossible" deps trees (from Julia's
perspective, that is) where we want to build something for Julia 1.6
that was compiled against GMP 6.1.2; although the ABIs are compatible,
Julia's resolver must be coerced to allow it.

This PR adds a `julia_version` keyword argument to all the methods from
`add()` down to the resolver.  `julia_version` defaults to `VERSION`,
but if it is given another version number, it resolves with respect to
that Julia version.  As the final word in any resolutionary argument, if
`julia_version` is set to `nothing`, then Julia compatibility bounds are
ignored completely, and only the `PackageSpec` bounds on other packaegs
are considered at all.

Usage of this faculty is not recommended for typical Julia package
installation.